Combined cpp files, stuffed helper files in hpp, updates start and game over screens (#2583)
