+++ firmware/application/external/breakout/main.cpp
@@ -20,13 +20,13 @@ void initialize_app(ui::NavigationView& nav) {
 extern "C" {
 
 __attribute__((section(".external_app.app_breakout.application_information"), used)) application_information_t _application_information_breakout = {
-    /*.memory_location = */ (uint8_t*)0x00000000,  // will be filled at compile time
-    /*.externalAppEntry = */ ui::external_app::breakout::initialize_app,
-    /*.header_version = */ CURRENT_HEADER_VERSION,
-    /*.app_version = */ VERSION_MD5,
+    (uint8_t*)0x00000000,
+    ui::external_app::breakout::initialize_app,
+    CURRENT_HEADER_VERSION,
+    VERSION_MD5,
 
-    /*.app_name = */ "Breakout",
-    /*.bitmap_data = */ {
+    "Breakout",
+    {
         0x00,
         0x00,
         0x7F,
@@ -60,11 +60,11 @@ __attribute__((section(".external_app.app_breakout.application_information"), us
         0xF7,
         0xF7,
     },
-    /*.icon_color = */ ui::Color::green().v,
-    /*.menu_location = */ app_location_t::GAMES,
-    /*.desired_menu_position = */ -1,
+    ui::Color::green().v,
+    app_location_t::GAMES,
+    -1,
 
-    /*.m4_app_tag = portapack::spi_flash::image_tag_none */ {0, 0, 0, 0},
-    /*.m4_app_offset = */ 0x00000000,  // will be filled at compile time
+    {0, 0, 0, 0},
+    0x00000000,
 };
-}
+}  // namespace ui::external_app::breakout
